+++ pokemon-app/src/hooks/index.js
@@ -1,0 +1,3 @@
+import useScroll from "./useScroll";
+
+export { useScroll };
